replaced GL Types with their C equivalent, change literal values to match their type for pedantic reasons

+++ src/core/OpenGL/VertexArrayBuffer.c
@@ -2,7 +2,7 @@
 
 void VertexArrayBuffer_init(VertexArrayBuffer *const vab, const void *const data, const unsigned int size){
     Lock_init(&vab->lock);
-    vab->index = 0;
+    vab->index = 0llu;
     glGenBuffers(1, &vab->id);
     glBindBuffer(GL_ARRAY_BUFFER, vab->id);
     glBufferData(GL_ARRAY_BUFFER, size, data, GL_DYNAMIC_DRAW);
@@ -11,7 +11,7 @@ void VertexArrayBuffer_init(VertexArrayBuffer *const vab, const void *const data
 inline void VertexArrayBuffer_push(VertexArrayBuffer *const vab, const void *const data, const unsigned int size) {
     glBindBuffer(GL_ARRAY_BUFFER, vab->id);
     glBufferSubData(GL_ARRAY_BUFFER, (long long int)vab->index, size, data);
-    vab->index += size;
+    vab->index += (size_t)size;
 }
 
 inline void VertexArrayBuffer_set(VertexArrayBuffer *const vab, const unsigned int index, const void *const data, const unsigned int size) {
@@ -30,5 +30,5 @@ inline void VertexArrayBuffer_bind(const VertexArrayBuffer *const vab) {
 
 inline void VertexArrayBuffer_unbind(const VertexArrayBuffer *const vab) {
     (void)vab;
-    glBindBuffer(GL_ARRAY_BUFFER, 0);
+    glBindBuffer(GL_ARRAY_BUFFER, 0u);
 }
